backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
flask==3.0.3
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
pycparser==2.22
werkzeug==3.0.2
zipp==3.18.1
